Modify type of threads

+++ erfurt.anna/M1/function.cpp
@@ -28,9 +28,9 @@ size_t erfurt::countPointsInCircul(c_it_t begin, c_it_t end)
   return std::accumulate(begin, end, 0);
 }
 
-double erfurt::getSquare(size_t seed, size_t numOfTests, double radius, unsigned int countOfThreads)
+double erfurt::getSquare(size_t seed, size_t numOfTests, double radius, size_t countOfThreads)
 {
-  size_t threads = std::min(std::thread::hardware_concurrency() - 1, countOfThreads);
+  size_t threads = std::min(static_cast<size_t>(std::thread::hardware_concurrency()) - 1, countOfThreads);
   std::vector< std::thread > ths;
   ths.reserve(threads);
   std::vector< size_t > results(threads, 0);
